refactor(core): Delete the redundant customFields seeding in ConfigService
`runPluginConfigurations` seeds an empty `customFields` array for every entity
that supports them, and it runs inside `preBootstrapConfig`, which completes
before the Nest app and so before any `ConfigService` exists. The CLI and
dashboard schema generators call `runPluginConfigurations` directly and never
build a `ConfigService`. Every key the getter would add is therefore already
present, so the getter now returns `activeConfig.customFields` unchanged.

That removes the second seeding site, and with it the need to keep two entity
list sources (`dbConnectionOptions.entities` and `getAllEntities(config)`) in
step. `config.service.spec.ts` covered the deleted loop; its one case that
tests the shared detection, a lookalike translation entity that is seeded,
moves to `bootstrap.spec.ts`. The metadata test helper now has a single
consumer, so it moves back into that spec file.

Relates to #4965
